middleware/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.31.2
h5py==3.6.0
idna==3.3
kiwisolver==1.4.2
logger==1.4
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandas-bokeh==0.5.5
Pillow==9.0.1
pvlib==0.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
scikit-learn==1.0.2
tornado==6.1
tornado-http-auth==1.1.1
urllib3==1.26.9
xgboost==1.5.1


